color and twitter message function update
