+++ constant.js
@@ -2,8 +2,4 @@ const yellowColor = "stop-color:#ffdb1e;stop-opacity:1;";
 const redColor = "stop-color:#ff0000;stop-opacity:1;";
 const greenColor = "stop-color:#6fcf0a;stop-opacity:1;";
 
-const treeValue = 3;
-const grassValue = 3;
-const weedsValue = 3;
-
-module.exports = {yellowColor,redColor,greenColor,treeValue,grassValue,weedsValue};
+module.exports = {yellowColor,redColor,greenColor};
